pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tkintertools-media"
authors = [{ name = "Xiaokang2022", email = "2951256653@qq.com" }]
maintainers = [{ name = "Xiaokang2022", email = "2951256653@qq.com" }]
description = "Extension package for tkintertools to matplotlib"
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">=3.10"
keywords = [
    "tkinter",
    "tkintertools",
    "media",
    "cross-platform",
    "themed",
    "video",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: Chinese (Simplified)",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "tkintertools>=3.0.0rc2",
    "pillow>=10.0.0",
    "ffpyplayer>=4.5.0",
    "typing_extensions",
]
dynamic = ["version"]

[project.urls]
"Bug tracker" = "https://github.com/Xiaokang2022/tkintertools-media/issues"
"Source code" = "https://github.com/Xiaokang2022/tkintertools-media"

[tool.setuptools.dynamic]
version = { attr = "tkintertools.media.__version__" }

[tool.setuptools.packages.find]
include = ["tkintertools*"]

# python -m build
# python -m twine upload dist/*
